Added project details card for new project

+++ lib/models/material_model.dart
@@ -1,9 +1,11 @@
+import 'dart:ffi';
+
 class MaterialModel {
   String name;
   String size;
   String type;
   String unit;
-  int price;
+  Float price;
   String image;
 
   MaterialModel({
